тз 13, попытка 2 добавлены интерфейсы, изменена логика работы с DTO сущностями
