+++ src/main/java/ru/practicum/shareit/request/ItemRequest.java
@@ -1,7 +1,7 @@
 package ru.practicum.shareit.request;
 
 import lombok.Data;
-import ru.practicum.shareit.user.User;
+import ru.practicum.shareit.user.model.User;
 
 import java.time.LocalDateTime;
