--- src/main/java/ru/practicum/shareit/user/UserDto.java
@@ -1,16 +1,0 @@
-package ru.practicum.shareit.user;
-
-import lombok.Data;
-
-@Data
-public class UserDto {
-
-    private String name;
-    private String email;
-
-    public UserDto(String name, String email) {
-        this.name = name;
-        this.email = email;
-    }
-
-}
