+++ src/main/java/ru/practicum/shareit/user/UserMapper.java
@@ -1,10 +1,21 @@
 package ru.practicum.shareit.user;
 
+import ru.practicum.shareit.user.dto.UserDto;
+import ru.practicum.shareit.user.model.User;
+
 public class UserMapper {
 
     public static UserDto toUserDto(User user) {
         return new UserDto(
+                user.getId(),
                 user.getName(),
                 user.getEmail());
+    }
+
+    public static User toUser(UserDto userDto) {
+        return new User(
+                userDto.getId(),
+                userDto.getName(),
+                userDto.getEmail());
     }
 }
