+++ src/main/java/ru/practicum/shareit/user/dto/UserDto.java
@@ -1,0 +1,12 @@
+package ru.practicum.shareit.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class UserDto {
+    private long id;
+    private String name;
+    private String email;
+}
